fix(mention-role): Fixes BOT-28

+++ src/commands/roles/mentionRole.ts
@@ -5,6 +5,8 @@ import { RoleResolver } from '../../resolvers';
 import { BotCommand, CommandGroup } from '../../types';
 import { Command, Context } from '../Command';
 
+const MISSING_PERMISSIONS = 50013;
+
 export default class extends Command {
 	public constructor(client: IMClient) {
 		super(client, {
@@ -53,7 +55,17 @@ export default class extends Command {
 				);
 			}
 
-			await role.edit({ mentionable: true }, 'Pinging role');
+			const res = await role
+				.edit({ mentionable: true }, 'Pinging role')
+				.catch(e => {
+					if (e.code === MISSING_PERMISSIONS) {
+						this.sendReply(message, 'cmd.mentionRole.missingPermissions');
+					}
+					return null as Role;
+				});
+			if (!res) {
+				return;
+			}
 			await message.channel.createMessage(`<@&${role.id}>`);
 			await role.edit({ mentionable: false }, 'Done pinging role');
 			await message.delete().catch(() => undefined);
